+++ richat/src/lib.rs
@@ -4,4 +4,5 @@ pub mod grpc;
 pub mod log;
 pub mod metrics;
 pub mod pubsub;
+pub mod source;
 pub mod version;