+++ CHANGELOG.md
@@ -16,7 +16,7 @@ The minor version will be incremented upon a breaking change and the patch versi
 
 ### Breaking
 
-## 2025-02-15
+## 2025-02-16
 
 - richat-cli-v2.2.1
 - richat-filter-v2.4.0
@@ -30,6 +30,7 @@ The minor version will be incremented upon a breaking change and the patch versi
 ### Breaking
 
 - filter: encode with slices ([#73](https://github.com/lamports-dev/richat/pull/73))
+- richat: remove parser thread ([#74](https://github.com/lamports-dev/richat/pull/74))
 
 ## 2025-02-11